+++ src/component/common/Checkbox.css
@@ -9,7 +9,7 @@
   appearance: none;
   width: 18px;
   height: 18px;
-  background-image: url("../image/icon/checkbox_0.png");
+  background-image: url("src/image/icon/checkbox_0.png");
   background-size: 100% 100%;
   margin-right: 7px;
   padding: 9px;
@@ -21,9 +21,9 @@
 }
 
 .checkbox:hover {
-  background-image: url("../image/icon/checkbox_1.png");
+  background-image: url("src/image/icon/checkbox_1.png");
 }
 
 .checkbox:checked {
-  background-image: url("../image/icon/checkbox.png");
+  background-image: url("src/image/icon/checkbox.png");
 }
